[universal-header] Remove unused style-utils
